chalice/pyproject.toml
[tool.poetry]
name = "bbprop_chalice"
version = "0.1.0"
description = ""
authors = ["Jeremy Morrison <jeremy.morrison36@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.17.41"
pandas = "^1.2.3"
requests = "^2.25.1"
python-dotenv = "^0.16.0"
chalice = "^1.22.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = "^20.8b1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
